requirements.txt
Django==1.7
django-jsonfield==0.9.13
djangorestframework==3.1.3
eventlog==0.11.0
wheel==0.24.0
